로그인 시 다른 Actor의 MsgActorPush 보내기 위한 사전작업

+++ Server/Source/Message/Message.cpp
@@ -22,5 +22,5 @@ MessageID Message::ID() const
 
 const int32_t Message::Size() const
 {
-	return m_body.size();
+	return (int32_t)(m_body.size());
 }
